move BaseFixtureGenerator to a new location
